silo-bridge/requirements.txt
slack-bolt>=1.18.0
anthropic>=0.39.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
